move test to wdio (not passing)
