--- test/karma/test-app/invisible-prehydration-false/index.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf8">
-<script src="./build/testinvisiblefalseprehydration.esm.js" type="module"></script>
-<script src="./build/testinvisiblefalseprehydration.js" nomodule></script>
-
-<prehydrated-styles></prehydrated-styles>
