--- test/karma/test-invisible-prehydration/.gitignore
@@ -1,1 +1,0 @@
-/test-www
