--- test/karma/test-invisible-prehydration/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "test-invisible-prehydration",
-  "main": "dist/index.cjs.js",
-  "module": "dist/index.js",
-  "collection": "dist/collection/collection-manifest.json",
-  "types": "dist/types/components.d.ts",
-  "volta": {
-    "extends": "../package.json"
-  }
-}
