--- test/karma/test-invisible-prehydration/tsconfig.json
@@ -1,33 +1,0 @@
-{
-  "compilerOptions": {
-    "alwaysStrict": true,
-    "allowSyntheticDefaultImports": true,
-    "allowUnreachableCode": true,
-    "declaration": false,
-    "resolveJsonModule": true,
-    "experimentalDecorators": true,
-    "forceConsistentCasingInFileNames": true,
-    "jsx": "react",
-    "jsxFactory": "h",
-    "lib": [
-      "dom",
-      "es2017"
-    ],
-    "module": "esnext",
-    "moduleResolution": "node",
-    "noImplicitAny": false,
-    "noImplicitReturns": false,
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "pretty": true,
-    "target": "es2017",
-    "useUnknownInCatchVariables": true,
-    "baseUrl": ".",
-    "paths": {
-      "@stencil/core": ["../../../internal"],
-      "@stencil/core/internal": ["../../../internal"],
-      "@stencil/core/testing": ["../../../testing"]
-    }
-  },
-  "include": ["src"]
-}
