+++ test/wdio/invisible-prehydration/components.d.ts
