+++ test/wdio/package.json
@@ -3,11 +3,12 @@
   "type": "module",
   "version": "0.0.0",
   "scripts": {
-    "build": "run-s build.global-script build.test-sibling build.main build.prerender",
+    "build": "run-s build.global-script build.test-sibling build.main build.prerender build.invisible-prehydration",
     "build.main": "node ../../bin/stencil build --debug --es5",
     "build.global-script": "node ../../bin/stencil build --debug --es5 --config global-script.stencil.config.ts",
     "build.test-sibling": "cd test-sibling && npm run build",
     "build.prerender": "node ../../bin/stencil build --config prerender.stencil.config.ts --prerender --debug && node ./test-prerender/prerender.js && node ./test-prerender/no-script-build.js",
+    "build.invisible-prehydration": "node ../../bin/stencil build --debug --es5 --config invisible-prehydration.stencil.config.ts",
     "test": "run-s build wdio",
     "wdio": "wdio run ./wdio.conf.ts"
   },
