+++ test/wdio/tsconfig-stencil.json
@@ -10,10 +10,7 @@
     "forceConsistentCasingInFileNames": true,
     "jsx": "react",
     "jsxFactory": "h",
-    "lib": [
-      "dom",
-      "es2017"
-    ],
+    "lib": ["dom", "es2017"],
     "module": "esnext",
     "moduleResolution": "node",
     "strictPropertyInitialization": false,
@@ -30,13 +27,7 @@
       "@test-sibling": ["./test-sibling"]
     }
   },
-  "include": [
-    "./src/components.d.ts",
-    "./**/*.spec.ts",
-    "./**/*.tsx",
-    "./util.ts",
-    "./global.ts"
-  ],
+  "include": ["./src/components.d.ts", "./**/*.spec.ts", "./**/*.tsx", "./util.ts", "./global.ts"],
   "exclude": [
     "./**/*.test.tsx",
     // we exclude the files in the global-script directory here since they're
@@ -44,6 +35,9 @@
     "./global-script/**/*.tsx",
     "./global-script/**/*.ts",
     // we also exclude the files in the test-sibling directory
-    "./test-sibling/**/*.tsx"
+    "./test-sibling/**/*.tsx",
+    // we also exclude the files in the invisible-prehydration directory
+    "./invisible-prehydration/**/*.tsx",
+    "./invisible-prehydration/**/*.ts"
   ]
 }
